Вынесет настройки путей и шапку в отдельные файлы
