+++ src/config.js
@@ -1,0 +1,8 @@
+const Routes = {
+  MAIN: `/`,
+  LOGIN: `/login`,
+  FAVORITES: `/favorites`,
+  HOTELS: `/hotels/:id`
+};
+
+export {Routes};
